feat: auth modal functionality * modularised *  in UI and auth restructured in server
