+++ client/src/constants/modalAssets.js
@@ -1,0 +1,9 @@
+export const addFormImg = {
+    src: 'https://www.publicbooks.org/wp-content/uploads/2022/05/arthur-mazi-a8CxRWIu8yw-unsplash-e1651170357711-810x522.jpg',
+    alt: 'Blue sky and rooftop'
+};
+
+export const editFormImg = {
+    src: 'https://w0.peakpx.com/wallpaper/281/219/HD-wallpaper-helicopter-sky-cloud-flight.jpg',
+    alt: 'Blue sky and helicopter'
+};
