+++ client/src/screens/Home.js
@@ -2,12 +2,12 @@ import Header from '../components/Header';
 import Hero from '../components/Hero';
 import Users from '../components/Users';
 import Footer from '../components/Footer';
-import AddProfileModal from '../components/AddUserModal';
+import Modal from '../components/Modal/Modal';
 
 const Home = () => {
     return (
         <main>
-            <AddProfileModal />
+            <Modal />
             <Header />
             <Hero />
             <Users />
