+++ client/src/styles/Users.css
@@ -21,6 +21,10 @@
 }
 
 @media screen and (min-width: 800px) {
+    .container {
+        padding: 20px 2.7%;
+    }
+
     #users-container {
         display: flex;
         flex-wrap: wrap;
@@ -28,7 +32,7 @@
 
     .container > h1 {
         font-size: 2em;
-    }  
+    }
 }
 
 .user-card {
@@ -37,8 +41,8 @@
     display: inline-block;
     margin-right: 3%;
     transition: all 0.2s ease;
-    animation-duration: 2s;
-    animation-delay: 1s;
+    animation-duration: 2.3s;
+    animation-delay: 0.75s;
     animation-timing-function: ease;
 }
 
@@ -82,6 +86,8 @@
 
 .user > img {
     width: 50%;
+    padding: 2%;
+    aspect-ratio: 1 / 1;
 }
 
 .user-textbox {
@@ -96,7 +102,7 @@
 @media screen and (min-width: 800px) {
     .user-textbox {
         position: absolute;
-        left: 50%;
+        left: 52%;
     }
 }
