+++ db/connection.js
@@ -7,14 +7,14 @@ class Database {
     this._connect()
   }
   
-_connect() {
-     mongoose.connect(uri)
-       .then(() => {
-         console.log('Database connection successful')
-       })
-       .catch(err => {
-         console.error(`Database connection error: ${err}`)
-       })
+  _connect() {
+    mongoose.connect(uri)
+      .then(() => {
+        console.log('Database connection successful')
+      })
+      .catch(err => {
+        console.error(`Database connection error: ${err}`)
+      })
   }
 }
