--- models/email.js
@@ -1,15 +1,0 @@
-const mongoose = require('mongoose');
-const Schema = mongoose.Schema;
-
-const EmailSchema = new Schema({
-    email: {
-        type: String,
-        trim: true,
-        lowercase: true,
-        unique: true,
-        required: true,
-        match: [/^\w+([\.-]?\w+)*@\w+([\.-]?\w+)*(\.\w{2,3})+$/, 'Please fill a valid email address']
-    }
-});
-
-module.exports = mongoose.model('Email', EmailSchema);
