+++ package.json
@@ -8,19 +8,23 @@
     "heroku-postbuild": "cd client && npm install && npm run build",
     "test": "echo \"Error: no test specified\" && exit 1",
     "client": "npm start --prefix client",
-    "dev": "concurrently \"node app.js\" \"npm run client\""
+    "dev": "concurrently \"nodemon app.js\" \"npm run client\""
   },
   "author": "",
   "license": "ISC",
   "dependencies": {
+    "bcrypt": "^5.0.1",
     "cloudinary": "^1.30.0",
     "concurrently": "^7.3.0",
     "cors": "^2.8.5",
     "dotenv": "^16.0.1",
     "express": "^4.18.1",
     "git": "^0.1.5",
+    "jsonwebtoken": "^8.5.1",
     "mongodb": "^4.7.0",
     "mongoose": "^6.3.6",
-    "node-fetch": "^2.6.1"
+    "mongoose-type-email": "^1.1.2",
+    "node-fetch": "^2.6.1",
+    "nodemon": "^2.0.19"
   }
 }
